fix: include shiki in output file tracing
