+++ .prettierignore
@@ -37,8 +37,5 @@ yarn-error.log*
 # vercel
 .vercel
 
-# shiki
-/src/content/mdx/shiki
-
 # typescript
 *.tsbuildinfo
