+++ package.json
@@ -19,11 +19,10 @@
     "dev": "next dev",
     "format": "prettier --config ./prettier.config.js --write \"**/*.{css,js,md,mjs,scss,ts,tsx}\"",
     "lint": "next lint",
-    "postinstall": "prisma generate && pnpm shiki:initialize",
+    "postinstall": "prisma generate",
     "prisma:generate": "prisma generate",
     "prisma:push": "prisma db push",
     "prisma:studio": "prisma studio -b none",
-    "shiki:initialize": "pnpm dlx mkdirp ./src/content/mdx/shiki && cp -r ./node_modules/shiki/{languages,themes} ./src/content/mdx/shiki/",
     "start": "next start",
     "typecheck": "tsc",
     "vscode:generate": "pnpm dlx tsx ./scripts/vscode.ts"
@@ -32,7 +31,7 @@
     "@headlessui/react": "1.7.11",
     "@hookform/resolvers": "2.9.11",
     "@next/bundle-analyzer": "13.1.6",
-    "@next/font": "13.1.6",
+    "@next/font": "13.1.7-canary.21",
     "@prisma/client": "4.10.1",
     "@tanstack/react-query": "4.24.6",
     "@tanstack/react-query-devtools": "4.24.6",
@@ -44,7 +43,7 @@
     "class-variance-authority": "0.4.0",
     "clsx": "1.2.1",
     "nanoid": "4.0.1",
-    "next": "13.1.6",
+    "next": "13.1.7-canary.21",
     "next-auth": "4.19.2",
     "next-mdx-remote": "4.3.0",
     "next-themes": "0.2.1",
