+++ src/content/mdx/serialize.ts
@@ -5,14 +5,44 @@ import {
   type Options as RehypeAutolinkHeadingsOptions,
 } from "rehype-autolink-headings";
 import rehypeKatex from "rehype-katex";
+import {
+  default as rehypePrettyCode,
+  type Options as RehypePrettyCodeOptions,
+} from "rehype-pretty-code";
 import rehypeSlug from "rehype-slug";
 import remarkGfm from "remark-gfm";
 import remarkMath from "remark-math";
 
-import { rehypeCode } from "./code";
+import dark from "./themes/dark.json";
+import light from "./themes/light.json";
+
+type RehypePrettyCodeNode = {
+  children: unknown[];
+  properties: {
+    className: string[];
+  };
+};
 
 const rehypeAutolinkHeadingsOptions: RehypeAutolinkHeadingsOptions = {
   behavior: "wrap",
+};
+
+const rehypePrettyCodeOptions: Partial<RehypePrettyCodeOptions> = {
+  onVisitHighlightedLine: (node: RehypePrettyCodeNode) => {
+    node.properties.className.push("highlighted");
+  },
+  onVisitHighlightedWord: (node: RehypePrettyCodeNode) => {
+    node.properties.className = ["word"];
+  },
+  onVisitLine: (node: RehypePrettyCodeNode) => {
+    if (node.children.length === 0) {
+      node.children = [{ type: "text", value: " " }];
+    }
+  },
+  theme: {
+    dark,
+    light,
+  } as unknown as Record<string, JSON>,
 };
 
 export const serializeMdx = (source: string) =>
@@ -21,7 +51,7 @@ export const serializeMdx = (source: string) =>
       format: "mdx",
       rehypePlugins: [
         rehypeKatex,
-        rehypeCode,
+        [rehypePrettyCode, rehypePrettyCodeOptions],
         rehypeSlug,
         [rehypeAutolinkHeadings, rehypeAutolinkHeadingsOptions],
       ],
